Make new subcommand files but it doesn't work
